[#20] test: LoginRequest 인스턴스 생성자 테스트 작성
