+++ src/main/java/com/study/realworld/domain/auth/dto/LoginRequest.java
@@ -1,0 +1,4 @@
+package com.study.realworld.domain.auth.dto;
+
+public class LoginRequest {
+}
