+++ src/main/java/com/study/realworld/domain/user/presentation/UserRestController.java
@@ -1,11 +1,10 @@
-package com.study.realworld.domain.user.ui;
+package com.study.realworld.domain.user.presentation;
 
 import com.study.realworld.domain.user.application.UserJoinService;
 import com.study.realworld.domain.user.domain.User;
 import com.study.realworld.domain.user.dto.UserJoinRequest;
 import com.study.realworld.domain.user.dto.UserJoinResponse;
 import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
