+++ src/test/java/com/study/realworld/domain/user/presentation/UserRestControllerUnitTest.java
@@ -1,16 +1,20 @@
-package com.study.realworld.domain.user.ui;
+package com.study.realworld.domain.user.presentation;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.study.realworld.domain.user.application.UserJoinService;
 import com.study.realworld.domain.user.domain.*;
 import com.study.realworld.domain.user.dto.UserJoinRequest;
+import com.study.realworld.domain.user.dto.UserJoinRequestTest;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
 import static com.study.realworld.domain.user.domain.BioTest.BIO;
 import static com.study.realworld.domain.user.domain.EmailTest.EMAIL;
@@ -19,37 +23,40 @@ import static com.study.realworld.domain.user.domain.NameTest.USERNAME;
 import static com.study.realworld.domain.user.domain.PasswordTest.PASSWORD;
 import static com.study.realworld.domain.user.domain.PasswordTest.PASSWORD_ENCODER;
 import static com.study.realworld.domain.user.domain.UserTest.userBuilder;
-import static com.study.realworld.domain.user.dto.UserJoinRequestTest.userJoinRequest;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
-import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-// 시큐리티 필터 제거
-@AutoConfigureMockMvc(addFilters = false)
-@WebMvcTest(UserRestController.class)
-class UserRestControllerTest {
+@ExtendWith(MockitoExtension.class)
+class UserRestControllerUnitTest {
 
-    @Autowired private MockMvc mockMvc;
-    @Autowired private ObjectMapper objectMapper;
-    @MockBean private UserJoinService userJoinService;
+    @Mock private UserJoinService userJoinService;
+    @InjectMocks private UserRestController userRestController;
 
-    @DisplayName("UserRestController 인스턴스의 join() 테스트")
+    private MockMvc mockMvc;
+    private ObjectMapper objectMapper;
+
+    @BeforeEach
+    void setUp() {
+        mockMvc = MockMvcBuilders.standaloneSetup(userRestController).build();
+        objectMapper = new ObjectMapper();
+    }
+
+    @DisplayName("UserRestController 인스턴스의 join() 단위 테스트")
     @Test
     void join_test() throws Exception {
-        final UserJoinRequest userJoinRequest = userJoinRequest(new Name(USERNAME), new Email(EMAIL), new Password(PASSWORD));
-        final User user = userBuilder(new Email(EMAIL), new Name(USERNAME), new Password(PASSWORD), new Bio(BIO), new Image(IMAGE)).encode(PASSWORD_ENCODER);
-        given(userJoinService.join(any())).willReturn(user);
-
+        final UserJoinRequest userJoinRequest = UserJoinRequestTest.userJoinRequest(new Name(USERNAME), new Email(EMAIL), new Password(PASSWORD));
         final String userJoinRequestString = objectMapper.writeValueAsString(userJoinRequest);
+        final User user = userBuilder(new Email(EMAIL), new Name(USERNAME), new Password(PASSWORD), new Bio(BIO), new Image(IMAGE)).encode(PASSWORD_ENCODER);
 
+        given(userJoinService.join(any())).willReturn(user);
         mockMvc.perform(post("/api/users")
-                .contentType(APPLICATION_JSON)
+                .contentType(MediaType.APPLICATION_JSON)
                 .content(userJoinRequestString))
                 .andExpect(status().isOk())
-                .andExpect(content().contentType(APPLICATION_JSON));
+                .andExpect(content().contentType(MediaType.APPLICATION_JSON));
     }
 
 }
